--- a/pkg/storage/pebble_merge.go
+++ b/pkg/storage/pebble_merge.go
@@ -7,13 +7,16 @@ package storage
 
 import (
 	"io"
+	"slices"
 	"sort"
+	"sync"
 
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage/enginepb"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/errors"
+	"github.com/cockroachdb/pebble"
 	"github.com/gogo/protobuf/proto"
 )
 
@@ -168,13 +171,20 @@ type MVCCValueMerger struct {
 	rawByteOps    [][]byte
 	oldestMergeTS hlc.LegacyTimestamp
 	oldToNew      bool
+	// inPool is set when we put this instance in the mvccValueMergerPool.
+	inPool bool
 
 	// Used to avoid heap allocations when passing pointer to `Unmarshal()`.
 	meta enginepb.MVCCMetadata
 	// merged and metaSubset are used to avoid heap allocations in Finish().
 	merged     roachpb.InternalTimeSeriesData
 	metaSubset enginepb.MVCCMetadataSubsetForMergeSerialization
+
+	// !!NOTE!! If any new fields are added to this struct, returnToPool needs to
+	// be updated accordingly.
 }
+
+var _ pebble.ValueMerger = (*MVCCValueMerger)(nil)
 
 const (
 	mvccChecksumSize = 4
@@ -182,43 +192,85 @@ const (
 	mvccHeaderSize   = mvccChecksumSize + 1
 )
 
+var mvccValueMergerPool = sync.Pool{
+	New: func() any {
+		return &MVCCValueMerger{}
+	},
+}
+
+func NewMVCCValueMerger() *MVCCValueMerger {
+	t := mvccValueMergerPool.Get().(*MVCCValueMerger)
+	t.inPool = false
+	return t
+}
+
+func (t *MVCCValueMerger) returnToPool() {
+	for i := range t.timeSeriesOps {
+		t.timeSeriesOps[i].ResetRetainingSlices()
+	}
+	t.timeSeriesOps = t.timeSeriesOps[:0]
+	t.rawByteOps = t.rawByteOps[:0]
+	t.oldestMergeTS = hlc.LegacyTimestamp{}
+	t.oldToNew = false
+
+	t.merged.ResetRetainingSlices()
+	rawBytesSlice := t.metaSubset.RawBytes[:0]
+	t.metaSubset = enginepb.MVCCMetadataSubsetForMergeSerialization{
+		RawBytes: rawBytesSlice,
+	}
+	t.resetMeta()
+	t.inPool = true
+	mvccValueMergerPool.Put(t)
+}
+
 func (t *MVCCValueMerger) ensureOrder(oldToNew bool) {
 	if oldToNew == t.oldToNew {
 		return
 	}
-	// Only one of the two loop bodies should actually execute under error-free
-	// conditions, i.e., all operands are either timeseries or all are non-
-	// timeseries.
-	for i := 0; i < len(t.timeSeriesOps)/2; i++ {
-		t.timeSeriesOps[i], t.timeSeriesOps[len(t.timeSeriesOps)-1-i] = t.timeSeriesOps[len(t.timeSeriesOps)-1-i], t.timeSeriesOps[i]
-	}
-	for i := 0; i < len(t.rawByteOps)/2; i++ {
-		t.rawByteOps[i], t.rawByteOps[len(t.rawByteOps)-1-i] = t.rawByteOps[len(t.rawByteOps)-1-i], t.rawByteOps[i]
-	}
+	// Only one of the two Reverse calls should actually do something under
+	// error-free conditions, i.e., all operands are either timeseries or all are
+	// non-timeseries.
+	slices.Reverse(t.timeSeriesOps)
+	slices.Reverse(t.rawByteOps)
 	t.oldToNew = oldToNew
 }
 
+func (t *MVCCValueMerger) resetMeta() {
+	rawBytes := t.meta.RawBytes[:0]
+	t.meta = enginepb.MVCCMetadata{RawBytes: rawBytes}
+}
+
 func (t *MVCCValueMerger) deserializeMVCCValueAndAppend(value []byte) error {
-	if err := protoutil.Unmarshal(value, &t.meta); err != nil {
+	t.resetMeta()
+	if err := t.meta.Unmarshal(value); err != nil { // nolint:protounmarshal
 		return errors.Wrap(err, "corrupted operand value")
 	}
 	if len(t.meta.RawBytes) < mvccHeaderSize {
 		return errors.Errorf("operand value too short")
 	}
 	if t.meta.RawBytes[mvccTagPos] == byte(roachpb.ValueType_TIMESERIES) {
-		if t.rawByteOps != nil {
+		if len(t.rawByteOps) > 0 {
 			return errors.Errorf("inconsistent value types for timeseries merge")
 		}
-		t.timeSeriesOps = append(t.timeSeriesOps, roachpb.InternalTimeSeriesData{})
+		if cap(t.timeSeriesOps) > len(t.timeSeriesOps) {
+			// Reuse any previous slices inside the InternalTimeSeriesData.
+			t.timeSeriesOps = t.timeSeriesOps[:len(t.timeSeriesOps)+1]
+		} else {
+			t.timeSeriesOps = append(t.timeSeriesOps, roachpb.InternalTimeSeriesData{})
+		}
 		ts := &t.timeSeriesOps[len(t.timeSeriesOps)-1]
-		if err := protoutil.Unmarshal(t.meta.RawBytes[mvccHeaderSize:], ts); err != nil {
+		// We want to reuse existing slices so we cannot use protoutil.Unmarshal().
+		if err := ts.Unmarshal(t.meta.RawBytes[mvccHeaderSize:]); err != nil { // nolint:protounmarshal
 			return errors.Wrap(err, "corrupted timeseries")
 		}
 	} else {
-		if t.timeSeriesOps != nil {
+		if len(t.timeSeriesOps) > 0 {
 			return errors.Errorf("inconsistent value types for non-timeseries merge")
 		}
-		t.rawByteOps = append(t.rawByteOps, t.meta.RawBytes[mvccHeaderSize:])
+		// Append to rawByteOps, reusing any slice that was already there.
+		n := len(t.rawByteOps)
+		t.rawByteOps = slices.Grow(t.rawByteOps, 1)[:n+1]
+		t.rawByteOps[n] = append(t.rawByteOps[n][:0], t.meta.RawBytes[mvccHeaderSize:]...)
 	}
 	// Save the timestamp of the oldest value since that is consistent with the
 	// behavior of the C++ DBMergeOperator.
@@ -232,6 +284,9 @@ func (t *MVCCValueMerger) deserializeMVCCValueAndAppend(value []byte) error {
 // (timeseries or non-timeseries). The slice will be reversed if needed such that it is in
 // old-to-new order.
 func (t *MVCCValueMerger) MergeNewer(value []byte) error {
+	if t.inPool {
+		return errors.AssertionFailedf("MVCCValueMerger used after being returned to pool")
+	}
 	t.ensureOrder(true /* oldToNew */)
 	if err := t.deserializeMVCCValueAndAppend(value); err != nil {
 		return err
@@ -243,6 +298,9 @@ func (t *MVCCValueMerger) MergeNewer(value []byte) error {
 // (timeseries or non-timeseries). The slice will be reversed if needed such that it is in
 // new-to-old order.
 func (t *MVCCValueMerger) MergeOlder(value []byte) error {
+	if t.inPool {
+		return errors.AssertionFailedf("MVCCValueMerger used after being returned to pool")
+	}
 	t.ensureOrder(false /* oldToNew */)
 	if err := t.deserializeMVCCValueAndAppend(value); err != nil {
 		return err
@@ -255,6 +313,10 @@ func (t *MVCCValueMerger) MergeOlder(value []byte) error {
 // of timeseries the values are sorted, deduplicated, and potentially migrated to columnar
 // format. When deduplicating, only the latest sample for a given offset is retained.
 func (t *MVCCValueMerger) Finish(includesBase bool) ([]byte, io.Closer, error) {
+	if t.inPool {
+		return nil, nil, errors.AssertionFailedf("MVCCValueMerger used after being returned to pool")
+	}
+	defer t.returnToPool()
 	isColumnar := false
 	if t.timeSeriesOps == nil && t.rawByteOps == nil {
 		return nil, nil, errors.Errorf("empty merge unsupported")
@@ -268,9 +330,8 @@ func (t *MVCCValueMerger) Finish(includesBase bool) ([]byte, io.Closer, error) {
 		}
 		// See the motivating comment in mvcc.proto.
 		meta := &t.metaSubset // avoid allocation
-		*meta = enginepb.MVCCMetadataSubsetForMergeSerialization{
-			RawBytes: make([]byte, mvccHeaderSize, mvccHeaderSize+totalLen),
-		}
+		meta.RawBytes = slices.Grow(meta.RawBytes[:0], mvccHeaderSize+totalLen)[:mvccHeaderSize]
+		clear(meta.RawBytes)
 		meta.RawBytes[mvccTagPos] = byte(roachpb.ValueType_BYTES)
 		for _, rawByteOp := range t.rawByteOps {
 			meta.RawBytes = append(meta.RawBytes, rawByteOp...)
@@ -287,11 +348,9 @@ func (t *MVCCValueMerger) Finish(includesBase bool) ([]byte, io.Closer, error) {
 	// compatible with any version that supports row format only. Then we can drop support
 	// for row format entirely. It requires significant cleanup effort as many tests target
 	// the row format.
-	merged := &t.merged // avoid allocation
-	*merged = roachpb.InternalTimeSeriesData{
-		StartTimestampNanos: t.timeSeriesOps[0].StartTimestampNanos,
-		SampleDurationNanos: t.timeSeriesOps[0].SampleDurationNanos,
-	}
+	merged := &t.merged // avoid allocation, reuse slices.
+	merged.StartTimestampNanos = t.timeSeriesOps[0].StartTimestampNanos
+	merged.SampleDurationNanos = t.timeSeriesOps[0].SampleDurationNanos
 	for i := range t.timeSeriesOps {
 		timeSeriesOp := &t.timeSeriesOps[i]
 		if timeSeriesOp.StartTimestampNanos != merged.StartTimestampNanos {
@@ -316,15 +375,14 @@ func (t *MVCCValueMerger) Finish(includesBase bool) ([]byte, io.Closer, error) {
 	}
 
 	meta := &t.metaSubset // avoid allocation
-	*meta = enginepb.MVCCMetadataSubsetForMergeSerialization{
-		RawBytes: make([]byte, mvccHeaderSize+merged.Size()),
-	}
+	rawBytesSize := mvccHeaderSize + merged.Size()
+	meta.RawBytes = slices.Grow(meta.RawBytes[:0], rawBytesSize)[:rawBytesSize]
 	meta.RawBytes[mvccTagPos] = byte(roachpb.ValueType_TIMESERIES)
 	// See the motivating comment in mvcc.proto.
 	if !(t.oldestMergeTS == hlc.LegacyTimestamp{}) {
 		meta.MergeTimestamp = &t.oldestMergeTS
 	}
-	_, err := protoutil.MarshalToSizedBuffer(merged, meta.RawBytes[mvccHeaderSize:])
+	_, err := merged.MarshalToSizedBuffer(meta.RawBytes[mvccHeaderSize:])
 	if err != nil {
 		return nil, nil, err
 	}