#### go.mod: bump gogoproto dep

Incorporate improvements to marshaling and unmarshaling.

Epic: none
Release note: None

#### storage: optimize MVCCValueMerger

We use a sync.Pool to speed up allocation of `MVCCValueMerger` and to
reuse internal slices. We use updated gogoproto marshalling and
unmarshaling code for which avoids unnecessary allocations and allows
reuse of slices.

```
name               old time/op    new time/op    delta
MVCCValueMerger-4    15.4µs ± 3%     7.6µs ± 1%  -50.67%  (p=0.000 n=8+8)

name               old alloc/op   new alloc/op   delta
MVCCValueMerger-4    86.2kB ± 0%    13.6kB ± 0%  -84.25%  (p=0.000 n=8+8)

name               old allocs/op  new allocs/op  delta
MVCCValueMerger-4      37.0 ± 0%       1.0 ± 0%  -97.30%  (p=0.000 n=8+8)
```

Informs: #133313
Release note: None