@RaduBerinde 2024-11-07
I will try to see if I can find a way that is cleaner than copy/pasting .pb.go code. Perhaps a patch to `internal.pg.go`. Unfortunately we don't have control over the marshalling code generation - I tried to switch to using our `gogoproto` fork but it was impossible without also forking a lot of dependencies which also use `gogoproto`.