@RaduBerinde 2024-11-08
*[`pkg/roachpb/internal_optimized_pb.go` line 11 at r9](https://reviewable.io/reviews/cockroachdb/cockroach/134481#-OBC8UlTCPuOh2BYyfQ9:-OBC8UlU9y7src-_v8vb:b-1h3cqj) ([raw file](https://github.com/cockroachdb/cockroach/blob/53fba58bd6045e76c54107612b16dd17c3bc1168/pkg/roachpb/internal_optimized_pb.go#L11)):*
> ```Go
> // InternalTimeSeriesData in internal.pb.go.
> //
> // The best way to look at the code in this file is to run `./dev gen protobuf`
> ```

Here is the diff from internal.pg.go: https://editor.mergely.com/bpbicztM


<!-- Sent from Reviewable.io -->