@RaduBerinde 2024-11-08
> Moving to draft for now. I will add functionality to protoc-gen-cockroach to apply a patch to an output file.

That turned to be way more messy. I moved the "semi-generated" code to its own file which is exempt from lint checks like `.pb.go` files are.