@jbowens 2024-11-08
should we consider moving away from protocol buffers in these low-level contexts (eg, something akin to flatbuffers)?