@RaduBerinde 2024-11-10
I think this is an ok solution for the time being. I will work on switching to our gogoproto fork (#129182) and incorporating these improvements so we generate better code.