@tbg 2024-11-12
@RaduBerinde I took the liberty to rework the commits in this PR so that the diffs in the optimized generated code could be reviewed (the original [diff](https://gist.github.com/tbg/4cf72c9a8cadac56f84b07b418a4d31b) couldn't meaningfully be inspected). There only diff between what I pushed and what you had is that I updated the comment at the top of `internal_optimized_pb.go` to say that the commit history has the diffs (vs. suggesting it could be inspected from diffing the files). Hope that was ok.