@tbg 2024-11-12
> should we consider moving away from protocol buffers in these low-level contexts (eg, something akin to flatbuffers)?

Yes, that would ultimately be desirable. However, we've been living for years with gogoproto and its unlikely that we'll manage to pull off a successful migration. It would have to be a priority, we'd need to support both for a while, and deal with the upgrades.
Raft traffic is another place where we'd prefer flatbuffers. Encoding overhead is very high in that area as well, and it doesn't help that we're using nested protobufs.

Forking gogoproto can potentially help us resolve the gridlock we're in w.r.t. protobuf upgrades, while also letting us make some very decent performance gains (as demonstrated here) without going all-in on the strategy in this PR.