@RaduBerinde 2024-11-12
Thank you Tobi! If we switch to the fork (https://github.com/cockroachdb/cockroach/pull/134918), I will move the optimizations into `gogoproto` so the generated code includes them by default. I will change this back to draft for now.