@RaduBerinde 2024-11-14
Updated to reuse `t.meta.RawBytes` as well (the `Unmarshal` generated code already tries to reuse it).